refactor: relocate gerber-view rendering from tabs into GerberViewPanel
