+++ crates/copperforge-core/src/ui/tabs.rs
@@ -1,23 +1,10 @@
 use crate::CopperForgeApp;
-use crate::layer_store::{mm_to_nm, nm_to_mm, mils_to_nm, nm_to_mils};
 use egui_citizen::message::CitizenId;
 
-use eframe::emath::{Rect, Vec2};
-use eframe::epaint::Color32;
-use egui::{Painter, Pos2, Stroke};
 use egui_dock::{SurfaceIndex, NodeIndex};
 use serde::{Serialize, Deserialize};
 
-use crate::event_logger::ReactiveEventLogger;
-use gerber_viewer::{
-    draw_crosshair,
-    draw_marker, ViewState
-};
-use crate::drc_operations::types::Position;
-use crate::display::manager::ToPosition;
-use nalgebra::Vector2;
 
-const MARKER_RADIUS: f32 = 6.0;
 
 /// Define the tabs for the DockArea
 #[derive(Clone, Serialize, Deserialize)]
@@ -76,7 +63,7 @@ pub struct Tab {
 
 impl Tab {
     /// Helper to get units from app
-    fn get_units(app: &CopperForgeApp) -> &crate::layer_store::UnitsState {
+    pub(crate) fn get_units(app: &CopperForgeApp) -> &crate::layer_store::UnitsState {
         &app.services.layer_store.units
     }
     
@@ -137,7 +124,7 @@ impl Tab {
                     .show(ui, params.app);
             }
             TabKind::GerberView => {
-                self.render_gerber_view(ui, params.app);
+                crate::panels::GerberViewPanel::new(egui_citizen::CitizenState::default()).render(ui, params.app);
             }
             TabKind::GerberView3d => {
                 // Rebuild meshes when the board geometry changed via any reload
@@ -186,1509 +173,6 @@ impl Tab {
             TabKind::Plugin(_) => unreachable!("handled before the match"),
         }
     }
-
-    fn render_gerber_view(&self, ui: &mut egui::Ui, app: &mut CopperForgeApp) {
-        // Render top controls
-        render_controls(ui, app);
-        ui.separator();
-        
-        // Set up viewport and handle interactions
-        let (viewport, response) = setup_viewport(ui, app);
-        handle_viewport_interactions(ui, app, &viewport, &response);
-        
-        // Render the gerber layers and overlays
-        render_gerber_content(ui, app, &viewport);
-    }
-}
-
-fn render_controls(ui: &mut egui::Ui, app: &mut CopperForgeApp) {
-    ui.vertical(|ui| {
-        // First row: PCB / gerber workflow (explicit actions, no auto-anything)
-        ui.horizontal(|ui| {
-            render_pcb_workflow_controls(ui, app);
-        });
-
-        ui.add_space(4.0);
-
-        // Second row: Main view controls
-        ui.horizontal(|ui| {
-            render_quadrant_controls(ui, app);
-            ui.separator();
-            render_layer_controls(ui, app);
-            ui.separator();
-            render_transform_controls(ui, app);
-        });
-
-        ui.add_space(4.0);
-
-        // Third row: Measurement and grid tools
-        ui.horizontal(|ui| {
-            render_zoom_display(ui, app);
-            ui.separator();
-            render_ruler_controls(ui, app);
-            ui.separator();
-            render_grid_controls(ui, app);
-        });
-    });
-}
-
-/// PCB file workflow on the Gerber Viewer ribbon — replaces the retired PCB File tab.
-fn render_pcb_workflow_controls(ui: &mut egui::Ui, app: &mut CopperForgeApp) {
-    use crate::project::{gerber_ops, ProjectState};
-
-    // Clone reactive handles up front so we can take `&mut app` later without
-    // a borrow conflict with a logger that borrows `app.services.logger_state`.
-    let logger_state = app.services.logger_state.clone();
-    let log_colors = app.services.log_colors.clone();
-    let logger = ReactiveEventLogger::with_colors(&logger_state, &log_colors);
-
-    // State label (dim when NoProject)
-    let (state_label, state_color) = match &app.services.project_state.get() {
-        ProjectState::NoProject => ("no PCB", egui::Color32::from_rgb(120, 120, 120)),
-        ProjectState::PcbSelected { .. } => ("PCB selected", egui::Color32::from_rgb(180, 180, 220)),
-        ProjectState::GeneratingGerbers { .. } => ("generating…", egui::Color32::from_rgb(230, 200, 120)),
-        ProjectState::GerbersGenerated { .. } => ("gerbers ready", egui::Color32::from_rgb(180, 220, 180)),
-        ProjectState::LoadingGerbers { .. } => ("loading…", egui::Color32::from_rgb(230, 200, 120)),
-        ProjectState::Ready { .. } => ("viewer loaded", egui::Color32::from_rgb(120, 220, 150)),
-    };
-    ui.label(egui::RichText::new(format!("● {state_label}")).color(state_color).monospace());
-
-    ui.separator();
-
-    let has_pcb = app.services.project_state.get().pcb_path().is_some();
-    let has_gerber_dir = app.services.project_state.get().gerber_dir().is_some();
-
-    // Generate: explicit. Disabled if no PCB is selected.
-    if ui.add_enabled(has_pcb, egui::Button::new("⚙ Generate Gerbers")).clicked() {
-        if let Some(pcb_path) = app.services.project_state.get().pcb_path().map(|p| p.to_path_buf()) {
-            let Some(method) = app.services.kicad_cli_method.clone() else {
-                logger.log_error("kicad-cli not found (checked PATH, Flatpak, and Snap at startup)");
-                return;
-            };
-            logger.log_info(&format!("Generating gerbers + drill via kicad-cli ({})…", method));
-            app.services.project_state.set(ProjectState::GeneratingGerbers { pcb_path: pcb_path.clone() });
-            if let Some(output_dir) = gerber_ops::generate_gerbers_from_pcb(&pcb_path, &method, &logger) {
-                app.services.project_state.set(ProjectState::GerbersGenerated { pcb_path, gerber_dir: output_dir });
-            } else {
-                app.services.project_state.set(ProjectState::PcbSelected { pcb_path });
-            }
-        }
-    }
-
-    // Load: disabled until gerbers have been generated (or already loaded — allows Reload).
-    let load_label = if matches!(app.services.project_state.get(), ProjectState::Ready { .. }) {
-        "↻ Reload Gerbers"
-    } else {
-        "⬇ Load Gerbers"
-    };
-    if ui.add_enabled(has_gerber_dir, egui::Button::new(load_label)).clicked() {
-        if let (Some(pcb_path), Some(gerber_dir)) = (
-            app.services.project_state.get().pcb_path().map(|p| p.to_path_buf()),
-            app.services.project_state.get().gerber_dir().map(|p| p.to_path_buf()),
-        ) {
-            app.services.project_state.set(ProjectState::LoadingGerbers { pcb_path: pcb_path.clone(), gerber_dir: gerber_dir.clone() });
-            gerber_ops::load_gerbers_into_viewer(app, &pcb_path, &gerber_dir, &logger);
-            let last_modified = std::fs::metadata(&pcb_path)
-                .and_then(|m| m.modified())
-                .unwrap_or(std::time::SystemTime::now());
-            app.services.project_state.set(ProjectState::Ready { pcb_path, gerber_dir, last_modified });
-        }
-    }
-
-    // Stale-file indicator
-    if let ProjectState::Ready { pcb_path, last_modified, .. } = &app.services.project_state.get() {
-        if let Ok(meta) = std::fs::metadata(pcb_path) {
-            if let Ok(modified) = meta.modified() {
-                if &modified != last_modified {
-                    ui.colored_label(egui::Color32::YELLOW, "⚠ PCB modified — regenerate");
-                }
-            }
-        }
-    }
-
-    ui.separator();
-
-    // Release: only enabled once gerbers are loaded AND a project record exists
-    // in the database (so the release can be persisted against a project).
-    let is_ready = matches!(app.services.project_state.get(), ProjectState::Ready { .. });
-    let has_current_project = app.projects_panel.panel_state.project_manager_state
-        .as_ref()
-        .and_then(|s| s.current_project.as_ref())
-        .is_some();
-    let release_enabled = is_ready && has_current_project;
-
-    let release_btn = ui.add_enabled(release_enabled, egui::Button::new("🚀 Release"));
-    let release_btn = if !has_current_project && is_ready {
-        release_btn.on_disabled_hover_text("Open or create a project record in the Projects tab first.")
-    } else if !is_ready {
-        release_btn.on_disabled_hover_text("Load gerbers (Generate + Load) before cutting a release.")
-    } else {
-        release_btn
-    };
-    if release_btn.clicked() {
-        open_release_modal(app);
-    }
-
-    // PCBWay-targeted release — identical pipeline + PCBWAY_FAB_SPECS.md
-    // (SMT/THT part + pad counts) added to the zip for the assembly quote.
-    let pcbway_btn = ui.add_enabled(release_enabled, egui::Button::new("🏭 Release for PCBWay"));
-    let pcbway_btn = if !has_current_project && is_ready {
-        pcbway_btn.on_disabled_hover_text("Open or create a project record in the Projects tab first.")
-    } else if !is_ready {
-        pcbway_btn.on_disabled_hover_text("Load gerbers (Generate + Load) before cutting a release.")
-    } else {
-        pcbway_btn.on_hover_text(
-            "Same as Release, but bundles PCBWAY_FAB_SPECS.md \
-             (SMT/THT part + pad counts, top/bottom split) into the zip.",
-        )
-    };
-    if pcbway_btn.clicked() {
-        open_release_modal_for(app, Some(crate::vendor::VendorKind::PcbWay));
-    }
-
-    // Pick up a pending "regenerate-release" intent from the Projects tab
-    // (set by right-click → Regenerate on a rev node). We open the release
-    // modal pre-filled with the existing rev's fields + overwrite flag.
-    let regen_tag = ui.ctx().memory(|mem| {
-        mem.data.get_temp::<String>(egui::Id::new("regen_release_intent"))
-    });
-    if let Some(composite) = regen_tag {
-        ui.ctx().memory_mut(|mem| {
-            mem.data.remove::<String>(egui::Id::new("regen_release_intent"));
-        });
-        open_regenerate_release_modal(app, &composite);
-    }
-
-    // Pick up a pending "load-release-gerbers" intent (set by right-click
-    // → Load Release Gerbers on a rev node). Extract the release ZIP into
-    // the user cache dir (skipped if already up-to-date) and load those
-    // gerbers into the viewer — no kicad-cli, no regenerate.
-    let load_tag = ui.ctx().memory(|mem| {
-        mem.data.get_temp::<String>(egui::Id::new("load_release_intent"))
-    });
-    if let Some(composite) = load_tag {
-        ui.ctx().memory_mut(|mem| {
-            mem.data.remove::<String>(egui::Id::new("load_release_intent"));
-        });
-        load_release_gerbers(app, &composite);
-    }
-
-    ui.separator();
-
-    if ui.add_enabled(has_pcb, egui::Button::new("✖ Clear")).clicked() {
-        app.services.project_state.set(ProjectState::NoProject);
-        if let Some(ref mut manager_state) = app.projects_panel.panel_state.project_manager_state {
-            manager_state.current_project = None;
-            manager_state.selected_project_id = None;
-        }
-        app.services.layer_store.clear_all();
-        logger.log_info("Cleared PCB file selection");
-    }
-}
-
-/// Seed the release modal with sensible defaults: next rev tag + current
-/// project description prefilled. Vendor-neutral (standard release).
-fn open_release_modal(app: &mut CopperForgeApp) {
-    open_release_modal_for(app, None);
-}
-
-/// Same as `open_release_modal` but targets a specific vendor — the
-/// modal's `target` field flows through `ReleaseRequest` into
-/// `create_release`, which injects the vendor-specific extras (e.g.
-/// `PCBWAY_FAB_SPECS.md`) into the zip.
-fn open_release_modal_for(app: &mut CopperForgeApp, target: Option<crate::vendor::VendorKind>) {
-    let (existing_releases, description_prefill, name_prefill) = app.projects_panel.panel_state.project_manager_state
-        .as_ref()
-        .and_then(|s| s.current_project.as_ref())
-        .map(|p| (p.releases.clone(), p.metadata.description.clone(), p.metadata.name.clone()))
-        .unwrap_or_default();
-    let suggested_tag = crate::release::suggest_next_rev_tag(&existing_releases);
-    app.projects_panel.panel_state.release_modal = Some(crate::app::ReleaseModalState {
-        rev_tag: suggested_tag,
-        description: description_prefill,
-        changes: String::new(),
-        board_pn: name_prefill,
-        copper_weight: String::new(),
-        include_date_in_name: true,
-        include_notes_in_zip: true,
-        error: None,
-        overwrite_existing: false,
-        target,
-    });
-}
-
-/// Resolve a `load_release` intent — parse the composite, look up the
-/// release archive + project PCB path, extract the ZIP (cached), and
-/// drive the existing gerber loader. Composite is "proj_X:rev:rev_02".
-fn load_release_gerbers(app: &mut CopperForgeApp, composite: &str) {
-    // Clone the reactive logger handles into owned locals so the logger
-    // doesn't keep a borrow on `app.services` for the whole function — that
-    // would collide with the `&mut app` passed to `load_gerbers_into_viewer`
-    // below. Same pattern the existing tabs.rs handlers use.
-    let logger_state = app.services.logger_state.clone();
-    let log_colors = app.services.log_colors.clone();
-    let logger = ReactiveEventLogger::with_colors(&logger_state, &log_colors);
-
-    let mut parts = composite.splitn(3, ':');
-    let project_id = match parts.next() { Some(s) => s.to_string(), None => return };
-    let _marker = parts.next();
-    let rev_tag = match parts.next() { Some(s) => s.to_string(), None => return };
-
-    // Resolve the release (carries archive_path).
-    let archive_path = app.projects_panel.panel_state.project_manager_state
-        .as_ref()
-        .and_then(|pm| pm.project_releases.get(&project_id))
-        .and_then(|releases| releases.iter().find(|r| r.tag == rev_tag))
-        .map(|r| r.archive_path.clone());
-    let Some(archive_path) = archive_path else {
-        logger.log_error(&format!(
-            "Load release: no release '{rev_tag}' on project '{project_id}'"
-        ));
-        return;
-    };
-
-    // Resolve the project's PCB path (load_gerbers_into_viewer needs it
-    // to read layer names from the .kicad_pcb stackup).
-    let pcb_path = app.projects_panel.panel_state.project_manager_state
-        .as_ref()
-        .and_then(|pm| pm.project_list.iter().find(|p| p.id == project_id))
-        .map(|p| p.pcb_file_path.clone());
-    let Some(pcb_path) = pcb_path else {
-        logger.log_error(&format!(
-            "Load release: no PCB file recorded for project '{project_id}'"
-        ));
-        return;
-    };
-
-    // Extract the ZIP's gerbers into the cache (skipped if already current).
-    let gerber_dir = match crate::release::extract_release_gerbers(&archive_path) {
-        Ok(dir) => {
-            logger.log_info(&format!(
-                "Extracted release '{rev_tag}' gerbers to {}",
-                dir.display()
-            ));
-            dir
-        }
-        Err(e) => {
-            logger.log_error(&format!(
-                "Load release: extraction of {} failed: {e}",
-                archive_path.display()
-            ));
-            return;
-        }
-    };
-
-    // Load the release gerbers into the viewer for INSPECTION ONLY. A release
-    // is a fab DELIVERABLE, not the project — so we deliberately do NOT touch
-    // `project_state` here. Leaving it on the actual project means a clean app
-    // close (which persists `project_state` via `save_settings` in `Drop`) saves
-    // the project, never the release's extracted gerbers. Previously this set
-    // `Ready { release_gerbers }`, which got persisted and wrongly restored as
-    // "the project" on the next launch.
-    crate::project::gerber_ops::load_gerbers_into_viewer(
-        app,
-        &pcb_path,
-        &gerber_dir,
-        &logger,
-    );
-    logger.log_info(&format!(
-        "Loaded gerbers from release '{rev_tag}' for inspection (project unchanged)"
-    ));
-}
-
-/// Open the release modal in Regenerate mode — seeded from the existing
-/// release's fields, overwrite flag set. `composite` is "proj_X:rev:rev_02".
-fn open_regenerate_release_modal(app: &mut CopperForgeApp, composite: &str) {
-    let mut parts = composite.splitn(3, ':');
-    let project_id = match parts.next() { Some(s) => s, None => return };
-    let _marker = parts.next();
-    let rev_tag = match parts.next() { Some(s) => s, None => return };
-
-    let existing = app.projects_panel.panel_state.project_manager_state
-        .as_ref()
-        .and_then(|pm| pm.project_releases.get(project_id))
-        .and_then(|releases| releases.iter().find(|r| r.tag == rev_tag))
-        .cloned();
-    let Some(existing) = existing else { return };
-
-    // board_pn/copper_weight aren't persisted on the Release record — prefill
-    // the part number from the project name; copper weight is re-entered.
-    let name_prefill = app.projects_panel.panel_state.project_manager_state
-        .as_ref()
-        .and_then(|pm| pm.project_list.iter().find(|p| p.id == project_id))
-        .map(|p| p.name.clone())
-        .unwrap_or_default();
-
-    app.projects_panel.panel_state.release_modal = Some(crate::app::ReleaseModalState {
-        rev_tag: existing.tag.clone(),
-        description: existing.description.clone(),
-        changes: existing.changes.clone(),
-        board_pn: name_prefill,
-        copper_weight: String::new(),
-        include_date_in_name: existing.include_date_in_name,
-        include_notes_in_zip: existing.include_notes_in_zip,
-        error: None,
-        overwrite_existing: true,
-        // Regenerate preserves the original vendor target — so re-cutting
-        // a PCBWay rev re-emits the PCBWay fab-specs sheet.
-        target: existing.target,
-    });
-}
-
-fn render_quadrant_controls(ui: &mut egui::Ui, app: &mut CopperForgeApp) {
-    if ui.checkbox(&mut app.services.display_manager.quadrant_view_enabled, "Quadrant View").clicked() {
-        app.services.layer_store.mark_dirty();
-        app.services.needs_initial_view = true;
-    }
-    
-    if app.services.display_manager.quadrant_view_enabled {
-        ui.separator();
-        ui.label("Spacing:");
-        
-        // Get units from ECS
-        let units_resource = Tab::get_units(app);
-        
-        let (mut spacing_value, units_suffix, conversion_factor) = if units_resource.is_mils() {
-            let spacing_nm = mm_to_nm(app.services.display_manager.quadrant_offset_magnitude);
-            (nm_to_mils(spacing_nm) as f32, "mils", 0.0254)
-        } else {
-            (app.services.display_manager.quadrant_offset_magnitude as f32, "mm", 1.0)
-        };
-
-        let speed = if units_resource.is_mils() { 10.0 } else { 1.0 };
-        let max_range = if units_resource.is_mils() { 20000.0 } else { 500.0 };
-        
-        if ui.add(egui::DragValue::new(&mut spacing_value)
-            .suffix(units_suffix)
-            .speed(speed)
-            .range(0.0..=max_range))
-            .changed() 
-        {
-            let spacing_mm = spacing_value * conversion_factor;
-            app.services.display_manager.set_quadrant_offset_magnitude(spacing_mm as f64);
-            app.services.layer_store.mark_dirty();
-        }
-        
-        ui.separator();
-        
-        if ui.button("📷 Export Layers as PNG").clicked() {
-            let logger_state = app.services.logger_state.clone();
-            let log_colors = app.services.log_colors.clone();
-            let logger = ReactiveEventLogger::with_colors(&logger_state, &log_colors);
-            crate::ui::orientation_panel::export_quadrant_layers_to_png(app, &logger);
-        }
-    }
-}
-
-fn render_layer_controls(ui: &mut egui::Ui, app: &mut CopperForgeApp) {
-    let flip_text = if app.services.display_manager.showing_top { "🔄 Flip to Bottom (F)" } else { "🔄 Flip to Top (F)" };
-    if ui.button(flip_text).clicked() {
-        app.services.display_manager.showing_top = !app.services.display_manager.showing_top;
-        
-        // Auto-toggle layer visibility based on flip state
-        use crate::layer_store::{LayerType, Side};
-        for layer_type in LayerType::all() {
-            let visible = match layer_type {
-                LayerType::Copper(1)
-                | LayerType::Silkscreen(Side::Top)
-                | LayerType::Soldermask(Side::Top)
-                | LayerType::Paste(Side::Top)
-                | LayerType::ViaPlugging(Side::Top) => {
-                    app.services.display_manager.showing_top
-                }
-                LayerType::Copper(_) => !app.services.display_manager.showing_top,
-                LayerType::Silkscreen(Side::Bottom)
-                | LayerType::Soldermask(Side::Bottom)
-                | LayerType::Paste(Side::Bottom)
-                | LayerType::ViaPlugging(Side::Bottom) => {
-                    !app.services.display_manager.showing_top
-                }
-                LayerType::MechanicalOutline | LayerType::Drill | LayerType::UserLayer(_) => {
-                    app.services.layer_store.get_visibility(layer_type)
-                }
-            };
-            app.services.layer_store.set_visibility(layer_type, visible);
-        }
-        
-        app.services.layer_store.mark_dirty();
-    }
-}
-
-fn render_transform_controls(ui: &mut egui::Ui, app: &mut CopperForgeApp) {
-    // Rotate button
-    if ui.button("🔄 Rotate (R)").clicked() {
-        app.services.rotation_degrees = (app.services.rotation_degrees + 90.0) % 360.0;
-        
-        // Don't reset view - just mark coordinates as dirty to update rotation
-        // This keeps the view centered on the current origin
-        app.services.layer_store.mark_dirty();
-        
-        let logger_state = app.services.logger_state.clone();
-        let log_colors = app.services.log_colors.clone();
-        let logger = ReactiveEventLogger::with_colors(&logger_state, &log_colors);
-        logger.log_custom(
-            crate::project::constants::LOG_TYPE_ROTATION, 
-            &format!("Rotated to {:.0}°", app.services.rotation_degrees)
-        );
-    }
-    
-    // Layer store rendering (gerber-viewer 0.2.0 compatible)
-    ui.label("🔥 Rendering (v0.2.0)");
-    
-    // Mirror buttons
-    let x_mirror_text = if app.services.display_manager.mirroring.x { "↔️ X Mirror ✓" } else { "↔️ X Mirror" };
-    if ui.button(x_mirror_text).clicked() {
-        app.services.display_manager.mirroring.x = !app.services.display_manager.mirroring.x;
-        // Don't reset custom origin, just mark coordinates as dirty
-        app.services.layer_store.mark_dirty();
-        
-        let logger_state = app.services.logger_state.clone();
-        let log_colors = app.services.log_colors.clone();
-        let logger = ReactiveEventLogger::with_colors(&logger_state, &log_colors);
-        logger.log_custom(
-            crate::project::constants::LOG_TYPE_MIRROR,
-            &format!("X mirroring {}", if app.services.display_manager.mirroring.x { "enabled" } else { "disabled" })
-        );
-    }
-    
-    let y_mirror_text = if app.services.display_manager.mirroring.y { "↕️ Y Mirror ✓" } else { "↕️ Y Mirror" };
-    if ui.button(y_mirror_text).clicked() {
-        app.services.display_manager.mirroring.y = !app.services.display_manager.mirroring.y;
-        // Don't reset custom origin, just mark coordinates as dirty
-        app.services.layer_store.mark_dirty();
-        
-        let logger_state = app.services.logger_state.clone();
-        let log_colors = app.services.log_colors.clone();
-        let logger = ReactiveEventLogger::with_colors(&logger_state, &log_colors);
-        logger.log_custom(
-            crate::project::constants::LOG_TYPE_MIRROR,
-            &format!("Y mirroring {}", if app.services.display_manager.mirroring.y { "enabled" } else { "disabled" })
-        );
-    }
-    
-    ui.separator();
-    
-    // Origin setting button
-    let origin_set = app.services.display_manager.design_offset.x != 0.0 || app.services.display_manager.design_offset.y != 0.0;
-    if origin_set {
-        if ui.button("🎯 Reset Origin").clicked() {
-            app.services.display_manager.design_offset = crate::display::VectorOffset { x: 0.0, y: 0.0 };
-            app.services.origin_has_been_set = false;
-            
-            // Force view refresh to properly center coordinates at the new origin
-            app.services.needs_initial_view = true;
-            
-            // Mark coordinates as dirty to force refresh
-            app.services.layer_store.mark_dirty();
-            
-            let logger_state = app.services.logger_state.clone();
-            let log_colors = app.services.log_colors.clone();
-            let logger = ReactiveEventLogger::with_colors(&logger_state, &log_colors);
-            logger.log_info("Reset origin to (0, 0) - view recentered");
-        }
-    } else {
-        if ui.button("🎯 Set Origin").clicked() {
-            app.services.setting_origin_mode = true;
-            
-            let logger_state = app.services.logger_state.clone();
-            let log_colors = app.services.log_colors.clone();
-            let logger = ReactiveEventLogger::with_colors(&logger_state, &log_colors);
-            logger.log_info("Click on the PCB to set the origin");
-        }
-    }
-}
-
-fn render_grid_controls(ui: &mut egui::Ui, app: &mut CopperForgeApp) {
-    ui.label("Grid:");
-    let grid_spacings_mils = [100.0, 50.0, 25.0, 10.0, 5.0, 2.0, 1.0];
-    let grid_spacings_mm = [2.54, 1.27, 0.635, 0.254, 0.127, 0.0508, 0.0254];
-    
-    let is_mils = {
-        let units_resource = Tab::get_units(app);
-        units_resource.is_mils()
-    };
-    
-    let spacings = if is_mils {
-        &grid_spacings_mils[..]
-    } else {
-        &grid_spacings_mm[..]
-    };
-    
-    // Find current selection
-    let mut current_spacing_display = "Custom".to_string();
-    for &spacing in spacings {
-        let spacing_mm = if is_mils {
-            nm_to_mm(mils_to_nm(spacing)) as f32
-        } else {
-            spacing as f32
-        };
-        if (app.services.grid_settings.spacing_mm - spacing_mm).abs() < 0.001 {
-            current_spacing_display = if is_mils {
-                format!("{} mils", spacing as i32)
-            } else {
-                format!("{:.3} mm", spacing)
-            };
-            break;
-        }
-    }
-    
-    egui::ComboBox::from_label("")
-        .selected_text(current_spacing_display)
-        .show_ui(ui, |ui| {
-            for &spacing in spacings {
-                let spacing_mm = if is_mils {
-                    nm_to_mm(mils_to_nm(spacing)) as f32
-                } else {
-                    spacing as f32
-                };
-                let label = if is_mils {
-                    format!("{} mils", spacing as i32)
-                } else {
-                    format!("{:.3} mm", spacing)
-                };
-                if ui.selectable_label(false, label).clicked() {
-                    app.services.grid_settings.spacing_mm = spacing_mm;
-                }
-            }
-        });
-    
-    ui.separator();
-    
-    // Grid dot size slider
-    ui.label("Dot Size:");
-    ui.add(egui::Slider::new(&mut app.services.grid_settings.dot_size, 0.5..=5.0).suffix("px"));
-    
-    ui.separator();
-    
-    // Enterprise feature: Snap to Grid
-    ui.checkbox(&mut app.services.grid_settings.snap_enabled, "🧲 Snap to Grid");
-}
-
-fn render_ruler_controls(ui: &mut egui::Ui, app: &mut CopperForgeApp) {
-    ui.label("📏 Ruler Tool:");
-    
-    let ruler_button_text = if app.services.ruler_active { "📏 Ruler ✓" } else { "📏 Ruler" };
-    if ui.button(ruler_button_text).clicked() {
-        app.services.ruler_active = !app.services.ruler_active;
-        if !app.services.ruler_active {
-            // Clear ruler when deactivated
-            app.services.ruler_start = None;
-            app.services.ruler_end = None;
-        }
-    }
-    
-    // Show ruler measurement if active and both points set
-    if app.services.ruler_active {
-        if let (Some(start), Some(end)) = (app.services.ruler_start, app.services.ruler_end) {
-            let dx = end.x - start.x;
-            let dy = end.y - start.y;
-            let distance = (dx * dx + dy * dy).sqrt();
-            
-            let units_resource = Tab::get_units(app);
-            if units_resource.is_mils() {
-                let distance_nm = mm_to_nm(distance);
-                let distance_mils = nm_to_mils(distance_nm);
-                let dx_nm = mm_to_nm(dx.abs());
-                let dy_nm = mm_to_nm(dy.abs());
-                ui.label(format!("📏 Distance: {:.2} mils", distance_mils));
-                ui.label(format!("📐 ΔX: {:.3} mils, ΔY: {:.3} mils", nm_to_mils(dx_nm), nm_to_mils(dy_nm)));
-            } else {
-                ui.label(format!("📏 Distance: {:.3} mm", distance));
-                ui.label(format!("📐 ΔX: {:.3} mm, ΔY: {:.3} mm", dx.abs(), dy.abs()));
-            }
-        } else if app.services.ruler_start.is_some() {
-            ui.label("Click second point to complete measurement");
-        } else {
-            ui.label("Click first point to start measurement (or press M to toggle)");
-        }
-    }
-    // Show latched measurement if not in active measurement mode
-    else if let (Some(start), Some(end)) = (app.services.latched_measurement_start, app.services.latched_measurement_end) {
-        let dx = end.x - start.x;
-        let dy = end.y - start.y;
-        let distance = (dx * dx + dy * dy).sqrt();
-        
-        let units_resource = Tab::get_units(app);
-        if units_resource.is_mils() {
-            let distance_nm = mm_to_nm(distance);
-            let distance_mils = nm_to_mils(distance_nm);
-            let dx_nm = mm_to_nm(dx.abs());
-            let dy_nm = mm_to_nm(dy.abs());
-            let dx_mils = nm_to_mils(dx_nm);
-            let dy_mils = nm_to_mils(dy_nm);
-            ui.label(egui::RichText::new(format!("📏 Distance: {:.2} mils", distance_mils)).color(egui::Color32::LIGHT_GRAY));
-            // Use more precision for deltas to avoid showing 0.00 for small values
-            ui.label(egui::RichText::new(format!("📐 ΔX: {:.3} mils, ΔY: {:.3} mils", dx_mils, dy_mils)).color(egui::Color32::LIGHT_GRAY));
-        } else {
-            ui.label(egui::RichText::new(format!("📏 Distance: {:.3} mm", distance)).color(egui::Color32::LIGHT_GRAY));
-            ui.label(egui::RichText::new(format!("📐 ΔX: {:.3} mm, ΔY: {:.3} mm", dx.abs(), dy.abs())).color(egui::Color32::LIGHT_GRAY));
-        }
-        ui.label(egui::RichText::new("(Previous measurement - press M to start new)").color(egui::Color32::GRAY).italics());
-    }
-}
-
-fn setup_viewport(ui: &mut egui::Ui, app: &mut CopperForgeApp) -> (Rect, egui::Response) {
-    // On-demand rendering. egui already repaints on input (pan / zoom / hover /
-    // keys), so we no longer force a redraw every frame — which previously
-    // re-rendered the whole (multi-up) scene at 60 fps even while idle, the main
-    // steady-state GPU/CPU drain. Keep repainting only while a view reset is
-    // pending, so a freshly loaded board still auto-fits without user input.
-    if app.services.needs_initial_view {
-        ui.ctx().request_repaint();
-    }
-
-    let available_size = ui.available_size();
-    let size = egui::Vec2::new(
-        available_size.x.max(100.0),
-        available_size.y.max(100.0)
-    );
-    
-    let response = ui.allocate_response(size, egui::Sense::click_and_drag());
-    let viewport = response.rect;
-    
-    // Handle double-click to center view (but maintain custom origin)
-    if response.double_clicked() {
-        // Only reset the view, don't change the custom origin (design_offset)
-        app.services.needs_initial_view = true;
-        
-        let logger_state = app.services.logger_state.clone();
-        let log_colors = app.services.log_colors.clone();
-        let logger = ReactiveEventLogger::with_colors(&logger_state, &log_colors);
-        logger.log_info("Centered view (double-click)");
-    }
-    
-    (viewport, response)
-}
-
-fn handle_viewport_interactions(ui: &mut egui::Ui, app: &mut CopperForgeApp, viewport: &Rect, response: &egui::Response) {
-    let mouse_pos_screen = ui.input(|i| i.pointer.hover_pos());
-
-    // KiCad mouse convention:
-    //   Left-drag  = zoom-to-area (rubber-band)
-    //   Right-drag = pan
-    //   Wheel      = zoom in/out
-    // Order matters: zoom-window runs first so its `zoom_window_dragging`
-    // flag gates the upstream gerber_viewer pan (which is hard-bound to
-    // left-button) out of the way during a rubber-band drag.
-    handle_zoom_window(ui, app, viewport, mouse_pos_screen, response);
-
-    // Handle mouse wheel zoom
-    handle_mouse_wheel_zoom(ui, app, viewport, response);
-
-    // Update UI state if not dragging zoom window
-    if !app.services.zoom_window_dragging {
-        // Local right-button pan (KiCad convention). Done before
-        // ui_state.update so the updated translation is what the rest of
-        // the frame sees. Upstream gerber_viewer's left-pan handler still
-        // fires from inside ui_state.update, but it's a no-op here because
-        // the user is dragging Secondary, not Primary.
-        handle_right_pan(ui, app, response);
-
-        app.services.ui_state.update(ui, viewport, response, &mut app.services.view_state);
-        
-        let viewport_center = viewport.center();
-        
-        // Calculate actual origin position based on custom origin if set
-        let design_offset = &app.services.display_manager.design_offset;
-        if design_offset.x != 0.0 || design_offset.y != 0.0 {
-            // Custom origin is set - convert to screen position
-            app.services.ui_state.origin_screen_pos = app.services.view_state.gerber_to_screen_coords(
-                Vector2::from(design_offset.clone()).to_position().to_point2()
-            );
-        } else {
-            // No custom origin - use viewport center
-            app.services.ui_state.origin_screen_pos = viewport_center;
-        }
-        
-        app.services.ui_state.center_screen_pos = viewport_center;
-        
-        // Update cursor coordinates using raw transform (not affected by design_offset)
-        if let Some(cursor_pos) = ui.input(|i| i.pointer.hover_pos()) {
-            // Use the original gerber coordinate system for origin setting
-            let raw_gerber_pos = app.services.view_state.screen_to_gerber_coords(cursor_pos);
-            app.services.ui_state.cursor_gerber_coords = Some(raw_gerber_pos);
-        }
-        
-        // Show visual feedback when in origin setting mode
-        if app.services.setting_origin_mode {
-            ui.ctx().set_cursor_icon(egui::CursorIcon::Crosshair);
-            
-            // Draw preview text at cursor
-            if let Some(mouse_pos) = ui.input(|i| i.pointer.hover_pos()) {
-                let painter = ui.painter();
-                painter.text(
-                    mouse_pos + Vec2::new(20.0, -20.0),
-                    egui::Align2::LEFT_BOTTOM,
-                    "Click to set origin",
-                    egui::FontId::default(),
-                    Color32::YELLOW,
-                );
-            }
-        }
-        
-        // Show visual feedback when in measurement mode
-        if app.services.ruler_active && !app.services.setting_origin_mode {
-            ui.ctx().set_cursor_icon(egui::CursorIcon::Crosshair);
-            
-            // Draw preview text at cursor
-            if let Some(mouse_pos) = ui.input(|i| i.pointer.hover_pos()) {
-                let painter = ui.painter();
-                
-                // Draw preview text
-                let text = if app.services.ruler_start.is_some() && app.services.ruler_end.is_none() {
-                    "Click second point to complete measurement"
-                } else {
-                    "Click to start measurement"
-                };
-                painter.text(
-                    mouse_pos + Vec2::new(20.0, -20.0),
-                    egui::Align2::LEFT_BOTTOM,
-                    text,
-                    egui::FontId::default(),
-                    Color32::CYAN,
-                );
-            }
-        }
-        
-        // Handle professional ruler tool with right-click drag
-        if app.services.ruler_active && !app.services.setting_origin_mode {
-            handle_ruler_interaction(ui, app, response);
-        }
-        
-        // Handle origin setting
-        if app.services.setting_origin_mode && response.clicked() {
-            if let Some(gerber_coords) = app.services.ui_state.cursor_gerber_coords {
-                // Enterprise feature: Apply snap to grid if enabled
-                let final_coords = if app.services.grid_settings.snap_enabled {
-                    let point = nalgebra::Point2::new(gerber_coords.x, gerber_coords.y);
-                    crate::display::snap_to_grid(point, &app.services.grid_settings)
-                } else {
-                    nalgebra::Point2::new(gerber_coords.x, gerber_coords.y)
-                };
-                
-                app.services.display_manager.design_offset = crate::display::VectorOffset {
-                    x: final_coords.x,
-                    y: final_coords.y,
-                };
-                app.services.setting_origin_mode = false;
-                app.services.origin_has_been_set = true;
-                
-                // Force view refresh to properly center coordinates at the new origin
-                app.services.needs_initial_view = true;
-                
-                // Mark coordinates as dirty to force refresh
-                app.services.layer_store.mark_dirty();
-                
-                let logger_state = app.services.logger_state.clone();
-                let log_colors = app.services.log_colors.clone();
-                let logger = ReactiveEventLogger::with_colors(&logger_state, &log_colors);
-                let snap_msg = if app.services.grid_settings.snap_enabled { " (snapped to grid)" } else { "" };
-                logger.log_info(&format!("Set origin to ({:.2}, {:.2}) mm{} - view recentered", final_coords.x, final_coords.y, snap_msg));
-            }
-        }
-    }
-}
-
-/// Pan the gerber view by right-mouse drag (KiCad convention).
-///
-/// Upstream `gerber_viewer::UiState::handle_panning` is hard-bound to the
-/// primary button — we can't reconfigure it, so this lives alongside as
-/// the right-button equivalent. Skipped while the ruler tool is active
-/// since ruler placement already binds right-drag.
-fn handle_right_pan(ui: &mut egui::Ui, app: &mut CopperForgeApp, response: &egui::Response) {
-    if app.services.ruler_active {
-        return;
-    }
-    if response.dragged_by(egui::PointerButton::Secondary) {
-        let delta = response.drag_delta();
-        if delta != Vec2::ZERO {
-            app.services.view_state.translation += delta;
-            ui.ctx().clear_animations();
-        }
-    }
-}
-
-fn handle_zoom_window(ui: &mut egui::Ui, app: &mut CopperForgeApp, viewport: &Rect, mouse_pos_screen: Option<Pos2>, response: &egui::Response) {
-    // KiCad: left-drag = rubber-band zoom-to-area. (Was Secondary before
-    // the right-drag-pans switch — see handle_viewport_interactions.)
-    let zoom_button = egui::PointerButton::Primary;
-
-    // Start zoom window
-    if response.contains_pointer() {
-        if ui.input(|i| i.pointer.button_pressed(zoom_button)) {
-            if let Some(pos) = mouse_pos_screen {
-                app.services.zoom_window_start = Some(pos);
-                app.services.zoom_window_dragging = true;
-            }
-        }
-    }
-
-    // Complete zoom window
-    if app.services.zoom_window_dragging && ui.input(|i| i.pointer.button_released(zoom_button)) {
-        if let (Some(start), Some(end)) = (app.services.zoom_window_start, ui.input(|i| i.pointer.hover_pos())) {
-            let zoom_rect = Rect::from_two_pos(start, end);
-            
-            if zoom_rect.width() > 10.0 && zoom_rect.height() > 10.0 {
-                let gerber_start = app.services.view_state.screen_to_gerber_coords(zoom_rect.min);
-                let gerber_end = app.services.view_state.screen_to_gerber_coords(zoom_rect.max);
-                
-                let gerber_width = (gerber_end.x - gerber_start.x).abs() as f32;
-                let gerber_height = (gerber_end.y - gerber_start.y).abs() as f32;
-                
-                let scale_x = viewport.width() / gerber_width;
-                let scale_y = viewport.height() / gerber_height;
-                let new_scale = scale_x.min(scale_y) * 0.9;
-                
-                let gerber_center_x = (gerber_start.x + gerber_end.x) / 2.0;
-                let gerber_center_y = (gerber_start.y + gerber_end.y) / 2.0;
-                
-                app.services.view_state.scale = new_scale;
-                
-                let viewport_center = viewport.center();
-                app.services.view_state.translation = Vec2::new(
-                    viewport_center.x - (gerber_center_x * new_scale as f64) as f32,
-                    viewport_center.y + (gerber_center_y * new_scale as f64) as f32
-                );
-                
-                app.sync_zoom_to_ecs(); // Sync zoom changes to ECS
-            }
-        }
-        
-        app.services.zoom_window_dragging = false;
-        app.services.zoom_window_start = None;
-    }
-    
-    // Cancel zoom window on escape
-    if app.services.zoom_window_dragging && ui.input(|i| i.key_pressed(egui::Key::Escape)) {
-        app.services.zoom_window_dragging = false;
-        app.services.zoom_window_start = None;
-    }
-}
-
-fn handle_mouse_wheel_zoom(ui: &mut egui::Ui, app: &mut CopperForgeApp, _viewport: &Rect, response: &egui::Response) {
-    if !response.contains_pointer() {
-        return;
-    }
-    
-    let scroll_delta = ui.input(|i| i.raw_scroll_delta.y);
-    if scroll_delta == 0.0 {
-        return;
-    }
-    
-    // Get mouse position for zoom center
-    let mouse_pos = ui.input(|i| i.pointer.hover_pos());
-    if let Some(mouse_pos) = mouse_pos {
-        // Calculate zoom factor (positive scroll = zoom in, negative = zoom out)
-        let zoom_factor = if scroll_delta > 0.0 {
-            1.1 // Zoom in
-        } else {
-            1.0 / 1.1 // Zoom out
-        };
-        
-        // Get current scale and calculate new scale
-        let old_scale = app.services.view_state.scale;
-        let new_scale = (old_scale * zoom_factor).clamp(0.01, 100.0);
-        
-        // Calculate the gerber coordinates at the mouse position before scaling
-        let gerber_point = app.services.view_state.screen_to_gerber_coords(mouse_pos);
-        
-        // Update the scale
-        app.services.view_state.scale = new_scale;
-        
-        // Calculate the new screen position of the same gerber point
-        let new_screen_pos = app.services.view_state.gerber_to_screen_coords(gerber_point);
-        
-        // Adjust translation to keep the mouse cursor over the same gerber point
-        let translation_adjustment = mouse_pos - new_screen_pos;
-        app.services.view_state.translation += translation_adjustment;
-        
-        // Sync with ECS zoom resource
-        app.sync_zoom_to_ecs();
-    }
-}
-
-fn render_gerber_content(ui: &mut egui::Ui, app: &mut CopperForgeApp, viewport: &Rect) {
-    let painter = ui.painter_at(*viewport);
-    painter.rect_filled(*viewport, 0.0, ui.visuals().extreme_bg_color);
-    
-    if app.services.needs_initial_view {
-        app.reset_view(*viewport);
-    }
-    
-    let painter = ui.painter().with_clip_rect(*viewport);
-    
-    // Draw grid
-    crate::display::draw_grid(&painter, viewport, &app.services.view_state, &app.services.grid_settings);
-
-    // Board substrate fill (behind the layers) so bare board — panel rails,
-    // routed gaps — reads as solid material rather than see-through background.
-    render_substrate_fill(app, &painter);
-
-    // Draw quadrant axes
-    if app.services.display_manager.quadrant_view_enabled {
-        draw_quadrant_axes(&painter, viewport, &app.services.view_state, app.services.ui_state.origin_screen_pos);
-    }
-    
-    // Draw crosshairs - always at the active origin
-    draw_crosshair(&painter, app.services.ui_state.origin_screen_pos, Color32::BLUE);
-    
-    // Render layers using ECS system (gerber-viewer 0.2.0 compatible)
-    app.render_layers_ecs(&painter);
-    
-    // Render overlays
-    render_overlays(app, &painter, viewport);
-
-    // Generic producer-supplied overlay (e.g. a dock panel's analysis shapes)
-    render_view_overlay(app, &painter);
-
-    // Render cursor info
-    render_cursor_info(ui, app, &painter, viewport);
-}
-
-/// Fill the board outline (Edge.Cuts) with an FR-4 colour behind the layers,
-/// so bare board reads as solid. Even-odd tessellation keeps routed slots /
-/// cutouts open. No-op when disabled or no board is loaded.
-fn render_substrate_fill(app: &CopperForgeApp, painter: &Painter) {
-    if !app.services.display_manager.show_substrate {
-        return;
-    }
-    let Some(outline) = app.services.board_outline.as_ref() else { return };
-    let Some((verts, indices)) = crate::gerber_geom::fill_mesh(&outline.contours) else {
-        return;
-    };
-    let vs = &app.services.view_state;
-    let color = Color32::from_rgb(28, 64, 40); // muted FR-4 green (substrate)
-    let vertices: Vec<egui::epaint::Vertex> = verts
-        .iter()
-        .map(|v| egui::epaint::Vertex {
-            pos: vs.gerber_to_screen_coords(nalgebra::Point2::new(v[0] as f64, v[1] as f64)),
-            uv: egui::epaint::WHITE_UV,
-            color,
-        })
-        .collect();
-    painter.add(egui::Shape::mesh(egui::epaint::Mesh {
-        indices,
-        vertices,
-        texture_id: egui::TextureId::default(),
-    }));
-}
-
-/// Draw overlays contributed by registered citizens (the viewport analogue of
-/// `register_panel`). Each plugin returns generic fills / lines / labels in
-/// world (mm) coords; core transforms and paints them without interpreting
-/// them. Core has no idea what any overlay represents.
-fn render_view_overlay(app: &CopperForgeApp, painter: &Painter) {
-    let vs = &app.services.view_state;
-    let pt = |p: (f64, f64)| vs.gerber_to_screen_coords(nalgebra::Point2::new(p.0, p.1));
-    let col = |c: [u8; 4]| Color32::from_rgba_unmultiplied(c[0], c[1], c[2], c[3]);
-
-    for panel in &app.plugin_panels {
-        let Some(ov) = panel.viewport_overlay(&app.services) else { continue };
-        for r in &ov.fills {
-            painter.rect_filled(egui::Rect::from_two_pos(pt(r.min), pt(r.max)), 0.0, col(r.rgba));
-        }
-        for l in &ov.lines {
-            painter.line_segment([pt(l.from), pt(l.to)], Stroke::new(l.width, col(l.rgba)));
-        }
-        for t in &ov.labels {
-            painter.text(
-                pt(t.at) + Vec2::new(4.0, 4.0),
-                egui::Align2::LEFT_TOP,
-                &t.text,
-                egui::FontId::proportional(t.size),
-                col(t.rgba),
-            );
-        }
-    }
-}
-
-
-fn render_overlays(app: &mut CopperForgeApp, painter: &Painter, viewport: &Rect) {
-    let screen_radius = MARKER_RADIUS * app.services.view_state.scale;
-    
-    // Origin marker - show only the active origin point
-    let design_offset = &app.services.display_manager.design_offset;
-    let has_custom_origin = design_offset.x != 0.0 || design_offset.y != 0.0;
-    
-    if has_custom_origin {
-        // Show custom origin (yellow marker) - this is the only visible origin
-        let design_offset_screen_position = app.services.view_state.gerber_to_screen_coords(Vector2::from(design_offset.clone()).to_position().to_point2());
-        draw_marker(painter, design_offset_screen_position, Color32::ORANGE, Color32::YELLOW, screen_radius);
-    } else {
-        // Show center origin (purple marker) when no custom origin is set
-        let purple_dot_pos = if app.services.display_manager.quadrant_view_enabled {
-            app.services.ui_state.center_screen_pos
-        } else {
-            app.services.ui_state.origin_screen_pos
-        };
-        draw_marker(painter, purple_dot_pos, Color32::PURPLE, Color32::MAGENTA, screen_radius);
-    }
-    
-    // Corner overlay shapes
-    render_corner_overlays(app, painter);
-    
-    // DRC violations
-    render_drc_violations(app, painter);
-    
-    // Board dimensions
-    render_board_dimensions(app, painter, viewport);
-    
-    // Enterprise feature: Ruler visualization
-    render_ruler(app, painter);
-    
-    // Custom measurement crosshair
-    render_measurement_crosshair(app, painter);
-    
-    // Zoom window
-    render_zoom_window(app, painter);
-}
-
-fn render_corner_overlays(app: &mut CopperForgeApp, painter: &Painter) {
-    if !app.services.drc_manager.corner_overlay_shapes.is_empty() {
-        let overlay_color = Color32::from_rgb(0, 255, 0);
-        
-        for shape in &app.services.drc_manager.corner_overlay_shapes {
-            let mut transformed_vertices = Vec::new();
-            
-            for point in &shape.points {
-                let mut vertex_pos = *point;
-                
-                // Apply rotation
-                if app.services.rotation_degrees != 0.0 {
-                    let rotation_radians = app.services.rotation_degrees.to_radians();
-                    let (sin_theta, cos_theta) = (rotation_radians.sin(), rotation_radians.cos());
-                    
-                    let rotated_x = vertex_pos.x * cos_theta as f64 - vertex_pos.y * sin_theta as f64;
-                    let rotated_y = vertex_pos.x * sin_theta as f64 + vertex_pos.y * cos_theta as f64;
-                    vertex_pos = Position { x: rotated_x, y: rotated_y };
-                }
-                
-                // Apply mirroring
-                if app.services.display_manager.mirroring.x {
-                    vertex_pos = vertex_pos.invert_x();
-                }
-                if app.services.display_manager.mirroring.y {
-                    vertex_pos = vertex_pos.invert_y();
-                }
-                
-                // Apply offsets
-                let origin = Vector2::from(app.services.display_manager.center_offset.clone()) - Vector2::from(app.services.display_manager.design_offset.clone());
-                vertex_pos = vertex_pos + origin.to_position();
-                
-                let vertex_screen = app.services.view_state.gerber_to_screen_coords(vertex_pos.to_point2());
-                transformed_vertices.push(vertex_screen);
-            }
-            
-            if transformed_vertices.len() >= 3 {
-                painter.add(egui::Shape::convex_polygon(
-                    transformed_vertices,
-                    overlay_color,
-                    Stroke::NONE
-                ));
-            }
-        }
-    }
-}
-
-fn render_drc_violations(app: &mut CopperForgeApp, painter: &Painter) {
-    for violation in &app.services.drc_manager.violations {
-        let violation_pos = Position::new(violation.x as f64, violation.y as f64);
-        let mut transformed_pos = violation_pos;
-        
-        // Apply rotation
-        if app.services.rotation_degrees != 0.0 {
-            let rotation_radians = app.services.rotation_degrees.to_radians();
-            let (sin_theta, cos_theta) = (rotation_radians.sin(), rotation_radians.cos());
-            let rotated_x = transformed_pos.x * cos_theta as f64 - transformed_pos.y * sin_theta as f64;
-            let rotated_y = transformed_pos.x * sin_theta as f64 + transformed_pos.y * cos_theta as f64;
-            transformed_pos = Position::new(rotated_x, rotated_y);
-        }
-        
-        // Apply mirroring
-        if app.services.display_manager.mirroring.x {
-            transformed_pos = transformed_pos.invert_x();
-        }
-        if app.services.display_manager.mirroring.y {
-            transformed_pos = transformed_pos.invert_y();
-        }
-        
-        // Apply offsets
-        let origin = Vector2::from(app.services.display_manager.center_offset.clone()) - Vector2::from(app.services.display_manager.design_offset.clone());
-        transformed_pos = transformed_pos + origin.to_position();
-        
-        let screen_pos = app.services.view_state.gerber_to_screen_coords(transformed_pos.to_point2());
-        
-        let base_size = 3.0;
-        let marker_size = base_size * app.services.view_state.scale.max(0.5);
-        let color = Color32::RED;
-        
-        draw_violation_marker(painter, screen_pos, marker_size, color);
-    }
-}
-
-fn render_board_dimensions(app: &mut CopperForgeApp, painter: &Painter, viewport: &Rect) {
-    if let Some(layer) = app.services.layer_store.find(crate::layer_store::LayerType::MechanicalOutline) {
-        let bbox = layer.gerber.bounding_box();
-        let width_mm = bbox.width();
-        let height_mm = bbox.height();
-        
-        let units_resource = Tab::get_units(app);
-        let dimension_text = if units_resource.is_mils() {
-            let width_nm = mm_to_nm(width_mm);
-            let height_nm = mm_to_nm(height_mm);
-            let width_mils = nm_to_mils(width_nm);
-            let height_mils = nm_to_mils(height_nm);
-            format!("{:.0} x {:.0} mils", width_mils, height_mils)
-        } else {
-            format!("{:.1} x {:.1} mm", width_mm, height_mm)
-        };
-        
-        let text_pos = viewport.max - Vec2::new(10.0, 50.0);
-        painter.text(
-            text_pos,
-            egui::Align2::RIGHT_BOTTOM,
-            dimension_text,
-            egui::FontId::default(),
-            Color32::from_rgb(200, 200, 200),
-        );
-    }
-}
-
-fn render_zoom_window(app: &mut CopperForgeApp, painter: &Painter) {
-    if app.services.zoom_window_dragging {
-        if let (Some(start), Some(current)) = (app.services.zoom_window_start, painter.ctx().input(|i| i.pointer.hover_pos())) {
-            let zoom_rect = Rect::from_two_pos(start, current);
-            
-            // Draw semi-transparent fill
-            painter.rect_filled(
-                zoom_rect,
-                0.0,
-                Color32::from_rgba_unmultiplied(100, 150, 255, 50)
-            );
-            
-            // Draw border
-            let stroke = Stroke::new(2.0, Color32::from_rgb(100, 150, 255));
-            painter.line_segment([zoom_rect.min, Pos2::new(zoom_rect.max.x, zoom_rect.min.y)], stroke);
-            painter.line_segment([Pos2::new(zoom_rect.max.x, zoom_rect.min.y), zoom_rect.max], stroke);
-            painter.line_segment([zoom_rect.max, Pos2::new(zoom_rect.min.x, zoom_rect.max.y)], stroke);
-            painter.line_segment([Pos2::new(zoom_rect.min.x, zoom_rect.max.y), zoom_rect.min], stroke);
-            
-            // Draw corner markers
-            let corner_size = 5.0;
-            let corners = [zoom_rect.min, 
-                          Pos2::new(zoom_rect.max.x, zoom_rect.min.y),
-                          zoom_rect.max,
-                          Pos2::new(zoom_rect.min.x, zoom_rect.max.y)];
-            
-            for corner in &corners {
-                painter.circle_filled(*corner, corner_size, Color32::from_rgb(100, 150, 255));
-            }
-        }
-    }
-}
-
-fn render_ruler(app: &mut CopperForgeApp, painter: &Painter) {
-    // Render active ruler if active
-    if app.services.ruler_active {
-        render_ruler_measurement(app, painter, app.services.ruler_start, app.services.ruler_end, true);
-    }
-    // Render latched ruler if not active but latched measurement exists
-    else if app.services.latched_measurement_start.is_some() && app.services.latched_measurement_end.is_some() {
-        render_ruler_measurement(app, painter, app.services.latched_measurement_start, app.services.latched_measurement_end, false);
-    }
-}
-
-fn render_ruler_measurement(app: &mut CopperForgeApp, painter: &Painter, start_opt: Option<nalgebra::Point2<f64>>, end_opt: Option<nalgebra::Point2<f64>>, is_active: bool) {
-    // Draw ruler points and line
-    if let Some(start) = start_opt {
-        let start_screen = app.services.view_state.gerber_to_screen_coords(start);
-        
-        // Choose colors based on active/latched state
-        let (point_color, line_color, text_color) = if is_active {
-            (Color32::RED, Color32::WHITE, Color32::WHITE)
-        } else {
-            (Color32::GRAY, Color32::LIGHT_GRAY, Color32::LIGHT_GRAY)
-        };
-        
-        // Draw start point
-        painter.circle_filled(start_screen, 4.0, point_color);
-        painter.circle_stroke(start_screen, 6.0, Stroke::new(2.0, line_color));
-        
-        if let Some(end) = end_opt {
-            let end_screen = app.services.view_state.gerber_to_screen_coords(end);
-            
-            // Draw end point
-            painter.circle_filled(end_screen, 4.0, point_color);
-            painter.circle_stroke(end_screen, 6.0, Stroke::new(2.0, line_color));
-            
-            // Draw ruler line
-            painter.line_segment(
-                [start_screen, end_screen],
-                Stroke::new(3.0, line_color)
-            );
-            
-            let dx = end.x - start.x;
-            let dy = end.y - start.y;
-            let distance = (dx * dx + dy * dy).sqrt();
-            
-            // Create measurement text with dx/dy display
-            let units_resource = Tab::get_units(app);
-            let measurement_text = if units_resource.is_mils() {
-                let distance_nm = mm_to_nm(distance);
-                let dx_nm = mm_to_nm(dx.abs());
-                let dy_nm = mm_to_nm(dy.abs());
-                format!(
-                    "{:.2} mils\nΔX: {:.2}\nΔY: {:.2}",
-                    nm_to_mils(distance_nm),
-                    nm_to_mils(dx_nm),
-                    nm_to_mils(dy_nm)
-                )
-            } else {
-                format!(
-                    "{:.3} mm\nΔX: {:.3}\nΔY: {:.3}",
-                    distance,
-                    dx,
-                    dy
-                )
-            };
-            
-            // Position text near the end point (offset to avoid overlap)
-            let text_offset = Vec2::new(20.0, -45.0);
-            let text_pos = end_screen + text_offset;
-            
-            // Draw text background (only for active measurements)
-            if is_active {
-                let text_size = painter.text(
-                    text_pos,
-                    egui::Align2::LEFT_TOP,
-                    "",
-                    egui::FontId::monospace(16.0),
-                    text_color,
-                ).size();
-                
-                let background_rect = egui::Rect::from_min_size(
-                    text_pos - Vec2::new(6.0, 6.0),
-                    text_size + Vec2::new(12.0, 12.0)
-                );
-                painter.rect_filled(background_rect, 6.0, Color32::from_rgba_unmultiplied(0, 0, 0, 240));
-            }
-            
-            // Draw measurement text at endpoint
-            painter.text(
-                text_pos,
-                egui::Align2::LEFT_TOP,
-                measurement_text,
-                egui::FontId::monospace(16.0),
-                text_color,
-            );
-        }
-    }
-}
-
-fn handle_ruler_interaction(ui: &mut egui::Ui, app: &mut CopperForgeApp, response: &egui::Response) {
-    if !app.services.ruler_active {
-        return;
-    }
-    
-    let mouse_pos = ui.input(|i| i.pointer.hover_pos());
-    
-    // In ruler mode, left-click to set measurement points
-    if response.clicked() {
-        if let Some(mouse_screen_pos) = mouse_pos {
-            let gerber_coords = app.services.view_state.screen_to_gerber_coords(mouse_screen_pos);
-            
-            // Apply snap to grid if enabled
-            let final_coords = if app.services.grid_settings.snap_enabled {
-                let point = nalgebra::Point2::new(gerber_coords.x, gerber_coords.y);
-                crate::display::snap_to_grid(point, &app.services.grid_settings)
-            } else {
-                nalgebra::Point2::new(gerber_coords.x, gerber_coords.y)
-            };
-            
-            if app.services.ruler_start.is_none() {
-                // First click - set start point
-                app.services.ruler_start = Some(final_coords);
-                app.services.ruler_end = None;
-                app.services.ruler_dragging = true; // Enable live preview
-            } else if app.services.ruler_end.is_none() {
-                // Second click - set end point and complete measurement
-                app.services.ruler_end = Some(final_coords);
-                app.services.ruler_dragging = false;
-            } else {
-                // Third click - start new measurement
-                app.services.ruler_start = Some(final_coords);
-                app.services.ruler_end = None;
-                app.services.ruler_dragging = true;
-            }
-        }
-    }
-    
-    // Show live preview when dragging (after first click, before second click)
-    if app.services.ruler_dragging && app.services.ruler_start.is_some() && mouse_pos.is_some() {
-        let mouse_screen_pos = mouse_pos.unwrap();
-        let gerber_coords = app.services.view_state.screen_to_gerber_coords(mouse_screen_pos);
-        
-        // Apply snap to grid if enabled
-        let final_coords = if app.services.grid_settings.snap_enabled {
-            let point = nalgebra::Point2::new(gerber_coords.x, gerber_coords.y);
-            crate::display::snap_to_grid(point, &app.services.grid_settings)
-        } else {
-            nalgebra::Point2::new(gerber_coords.x, gerber_coords.y)
-        };
-        
-        // Update live preview end point
-        app.services.ruler_end = Some(final_coords);
-    }
-}
-
-fn render_cursor_info(ui: &mut egui::Ui, app: &mut CopperForgeApp, painter: &Painter, viewport: &Rect) {
-    // Hide cursor coordinates when ruler mode is active
-    if app.services.ruler_active {
-        return;
-    }
-    
-    let mouse_pos_screen = ui.input(|i| i.pointer.hover_pos());
-    
-    if let Some(mouse_screen_pos) = mouse_pos_screen {
-        if viewport.contains(mouse_screen_pos) {
-            let gerber_pos = app.services.view_state.screen_to_gerber_coords(mouse_screen_pos);
-            
-            // Apply the design_offset as a simple coordinate offset for display
-            // The design_offset is where we want (0,0) to be, so we subtract it from current position
-            let adjusted_pos = Position::new(
-                gerber_pos.x - app.services.display_manager.design_offset.x,
-                gerber_pos.y - app.services.display_manager.design_offset.y
-            );
-            
-            let units_resource = Tab::get_units(app);
-            let cursor_text = if units_resource.is_mils() {
-                let x_nm = mm_to_nm(adjusted_pos.x as f64);
-                let y_nm = mm_to_nm(adjusted_pos.y as f64);
-                let x_mils = nm_to_mils(x_nm);
-                let y_mils = nm_to_mils(y_nm);
-                format!("({:.0}, {:.0}) mils", x_mils, y_mils)
-            } else {
-                format!("({:.2}, {:.2}) mm", adjusted_pos.x, adjusted_pos.y)
-            };
-            
-            
-            let text_offset = Vec2::new(15.0, -15.0);
-            let cursor_text_pos = mouse_screen_pos + text_offset;
-            
-            // Draw background
-            let text_size = painter.text(
-                cursor_text_pos,
-                egui::Align2::LEFT_TOP,
-                "",
-                egui::FontId::monospace(12.0),
-                Color32::WHITE,
-            ).size();
-            
-            let background_rect = egui::Rect::from_min_size(
-                cursor_text_pos - Vec2::new(2.0, 2.0),
-                text_size + Vec2::new(4.0, 4.0)
-            );
-            
-            painter.rect_filled(
-                background_rect,
-                3.0,
-                Color32::from_rgba_unmultiplied(0, 0, 0, 180)
-            );
-            
-            // Draw text
-            painter.text(
-                cursor_text_pos,
-                egui::Align2::LEFT_TOP,
-                cursor_text,
-                egui::FontId::monospace(12.0),
-                Color32::WHITE,
-            );
-            
-            // Draw crosshair
-            let crosshair_size = 8.0;
-            painter.line_segment(
-                [
-                    mouse_screen_pos - Vec2::new(crosshair_size, 0.0),
-                    mouse_screen_pos + Vec2::new(crosshair_size, 0.0)
-                ],
-                Stroke::new(1.0, Color32::WHITE)
-            );
-            painter.line_segment(
-                [
-                    mouse_screen_pos - Vec2::new(0.0, crosshair_size),
-                    mouse_screen_pos + Vec2::new(0.0, crosshair_size)
-                ],
-                Stroke::new(1.0, Color32::WHITE)
-            );
-        }
-    }
-    
-    // Unit display
-    let unit_toggle_pos = viewport.max - Vec2::new(10.0, 30.0);
-    let units_resource = Tab::get_units(app);
-    let unit_text = if units_resource.is_mils() { "mils" } else { "mm" };
-    painter.text(
-        unit_toggle_pos,
-        egui::Align2::RIGHT_BOTTOM,
-        format!("Mouse: {}", unit_text),
-        egui::FontId::default(),
-        Color32::from_rgb(150, 150, 150),
-    );
-}
-
-fn render_measurement_crosshair(app: &mut CopperForgeApp, painter: &Painter) {
-    // Skip if in origin setting mode
-    if app.services.setting_origin_mode {
-        return;
-    }
-    
-    // Draw crosshairs for active measurement points
-    if app.services.ruler_active {
-        if let Some(start_point) = app.services.ruler_start {
-            let screen_pos = app.services.view_state.gerber_to_screen_coords(start_point);
-            draw_measurement_crosshair(painter, screen_pos, Color32::from_rgb(139, 0, 0)); // Dark red for active
-        }
-        
-        if let Some(end_point) = app.services.ruler_end {
-            let screen_pos = app.services.view_state.gerber_to_screen_coords(end_point);
-            draw_measurement_crosshair(painter, screen_pos, Color32::from_rgb(139, 0, 0)); // Dark red for active
-        }
-    }
-    // Draw crosshairs for latched measurement points (grayed out)
-    else {
-        if let Some(start_point) = app.services.latched_measurement_start {
-            let screen_pos = app.services.view_state.gerber_to_screen_coords(start_point);
-            draw_measurement_crosshair(painter, screen_pos, Color32::from_rgb(100, 100, 100)); // Gray for latched
-        }
-        
-        if let Some(end_point) = app.services.latched_measurement_end {
-            let screen_pos = app.services.view_state.gerber_to_screen_coords(end_point);
-            draw_measurement_crosshair(painter, screen_pos, Color32::from_rgb(100, 100, 100)); // Gray for latched
-        }
-    }
-}
-
-fn draw_measurement_crosshair(painter: &Painter, center: Pos2, color: Color32) {
-    let crosshair_size = 12.0;
-    let stroke = egui::Stroke::new(2.0, color);
-    
-    // Horizontal line
-    painter.line_segment(
-        [
-            center + Vec2::new(-crosshair_size, 0.0),
-            center + Vec2::new(crosshair_size, 0.0),
-        ],
-        stroke,
-    );
-    
-    // Vertical line
-    painter.line_segment(
-        [
-            center + Vec2::new(0.0, -crosshair_size),
-            center + Vec2::new(0.0, crosshair_size),
-        ],
-        stroke,
-    );
 }
 
 pub struct TabViewer<'a> {
@@ -1721,77 +205,4 @@ impl<'a> egui_dock::TabViewer for TabViewer<'a> {
         };
         tab.content(ui, &mut params);
     }
-}
-
-fn render_zoom_display(ui: &mut egui::Ui, app: &mut CopperForgeApp) {
-    // Get zoom info from ECS, fallback to legacy ViewState
-    let (zoom_percentage, scale_factor) = (
-        app.services.layer_store.zoom.zoom_percentage(),
-        app.services.layer_store.zoom.scale,
-    );
-    
-    // Format zoom display with appropriate precision
-    let zoom_text = if zoom_percentage >= 100.0 {
-        format!("🔍 {:.0}%", zoom_percentage)
-    } else if zoom_percentage >= 10.0 {
-        format!("🔍 {:.1}%", zoom_percentage)
-    } else {
-        format!("🔍 {:.2}%", zoom_percentage)
-    };
-    
-    // Display zoom with a distinct visual style
-    ui.label(egui::RichText::new(zoom_text)
-        .color(egui::Color32::from_rgb(100, 200, 100))
-        .strong())
-        .on_hover_text(format!(
-            "Current Zoom Level (ECS)\nScale Factor: {:.3}x\nPercentage: {:.2}%",
-            scale_factor,
-            zoom_percentage
-        ));
-}
-
-/// Draw a red X marker for DRC violations
-fn draw_violation_marker(painter: &Painter, center: Pos2, size: f32, color: Color32) {
-    let half_size = size / 2.0;
-    let stroke = Stroke::new(2.0, color);
-    
-    // Draw X shape
-    painter.line_segment([
-        Pos2::new(center.x - half_size, center.y - half_size),
-        Pos2::new(center.x + half_size, center.y + half_size)
-    ], stroke);
-    
-    painter.line_segment([
-        Pos2::new(center.x - half_size, center.y + half_size),
-        Pos2::new(center.x + half_size, center.y - half_size)
-    ], stroke);
-}
-
-/// Draw quadrant axes when quadrant view is enabled
-fn draw_quadrant_axes(painter: &Painter, viewport: &Rect, _view_state: &ViewState, center_screen_pos: Pos2) {
-    let stroke = Stroke::new(2.0, Color32::from_rgba_unmultiplied(100, 100, 100, 150));
-    
-    // Draw vertical axis
-    if center_screen_pos.x >= viewport.min.x && center_screen_pos.x <= viewport.max.x {
-        painter.line_segment(
-            [
-                Pos2::new(center_screen_pos.x, viewport.min.y),
-                Pos2::new(center_screen_pos.x, viewport.max.y)
-            ],
-            stroke
-        );
-    }
-    
-    // Draw horizontal axis
-    if center_screen_pos.y >= viewport.min.y && center_screen_pos.y <= viewport.max.y {
-        painter.line_segment(
-            [
-                Pos2::new(viewport.min.x, center_screen_pos.y),
-                Pos2::new(viewport.max.x, center_screen_pos.y)
-            ],
-            stroke
-        );
-    }
-    
-    // Quadrant labels removed as requested by user
 }
